Move Login button to own directory
