+++ src/components/Login/Login.less
@@ -1,0 +1,5 @@
+.login-button-wrapper {
+  align-self: center;
+  position: relative;
+  left: 200px;
+}
